+++ next-docs/public/examples/tooltipTW/Default.tsx
@@ -1,0 +1,15 @@
+import { Tooltip, Chip } from '@heathmont/moon-core-tw';
+
+const Example = () => (
+  <Tooltip>
+    <Tooltip.Trigger>
+      <Chip>Trigger</Chip>
+    </Tooltip.Trigger>
+    <Tooltip.Content>
+      This is the default tooltip
+      <Tooltip.Arrow />
+    </Tooltip.Content>
+  </Tooltip>
+);
+
+export default Example;