+++ next-docs/public/examples/tooltipTW/__tests__/index.test.tsx
@@ -1,0 +1,109 @@
+/**
+ * @jest-environment jsdom
+ */
+
+import React from 'react';
+import renderer from 'react-test-renderer';
+import {
+  moonDesignDark,
+  moonDesignLight,
+  ThemeProvider,
+} from '@heathmont/moon-themes';
+import ArrowPositions from '../ArrowPositions';
+import Default from '../Default';
+import Customization from '../Customization';
+import WithoutArrow from '../WithoutArrow';
+import WithoutShadow from '../WithoutShadow';
+
+const renderWithLightTheme = (component: JSX.Element) => (
+  <ThemeProvider theme={moonDesignLight}>{component}</ThemeProvider>
+);
+const renderWithDarkTheme = (component: JSX.Element) => (
+  <ThemeProvider theme={moonDesignDark}>{component}</ThemeProvider>
+);
+const renderWithRtl = (component: JSX.Element) => (
+  <div dir="rtl">{component}</div>
+);
+
+describe('TextInput in Light Theme', () => {
+  it('renders Default', () => {
+    const testRenderer = renderer.create(renderWithLightTheme(<Default />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+  it('renders ArrowPositions', () => {
+    const testRenderer = renderer.create(
+      renderWithLightTheme(<ArrowPositions />)
+    );
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+  it('renders Customization', () => {
+    const testRenderer = renderer.create(
+      renderWithLightTheme(<Customization />)
+    );
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+  it('renders WithoutArrow', () => {
+    const testRenderer = renderer.create(
+      renderWithLightTheme(<WithoutArrow />)
+    );
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+  it('renders WithoutShadow', () => {
+    const testRenderer = renderer.create(
+      renderWithLightTheme(<WithoutShadow />)
+    );
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+});
+
+describe('TextInput in Dark Theme', () => {
+  it('renders Default', () => {
+    const testRenderer = renderer.create(renderWithDarkTheme(<Default />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+  it('renders ArrowPositions', () => {
+    const testRenderer = renderer.create(
+      renderWithDarkTheme(<ArrowPositions />)
+    );
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+  it('renders Customization', () => {
+    const testRenderer = renderer.create(
+      renderWithDarkTheme(<Customization />)
+    );
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+  it('renders WithoutArrow', () => {
+    const testRenderer = renderer.create(renderWithDarkTheme(<WithoutArrow />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+  it('renders WithoutShadow', () => {
+    const testRenderer = renderer.create(
+      renderWithDarkTheme(<WithoutShadow />)
+    );
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+});
+
+describe('TextInput in RTL', () => {
+  it('renders Default', () => {
+    const testRenderer = renderer.create(renderWithRtl(<Default />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+  it('renders ArrowPositions', () => {
+    const testRenderer = renderer.create(renderWithRtl(<ArrowPositions />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+  it('renders Customization', () => {
+    const testRenderer = renderer.create(renderWithRtl(<Customization />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+  it('renders WithoutArrow', () => {
+    const testRenderer = renderer.create(renderWithRtl(<WithoutArrow />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+  it('renders WithoutShadow', () => {
+    const testRenderer = renderer.create(renderWithRtl(<WithoutShadow />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+});