+++ workspaces/core/src/index.ts
@@ -54,5 +54,7 @@ export { default as TextInput } from './textInput/TextInput';
 export * from './textInput/TextInput';
 export { default as TextInputGroup } from './textInputGroup/TextInputGroup';
 export * from './textInputGroup/TextInputGroup';
+export { default as Tooltip } from './tooltip/Tooltip';
+export * from './tooltip/Tooltip';
 export { default as mergeClassnames } from './utils/mergeClassnames';
 export * from './utils/mergeClassnames';